+++ KSPAddon.cs
@@ -12,17 +12,19 @@ namespace KRPC.MechJeb {
 
 	[KSPAddon(KSPAddon.Startup.Flight, false)]
 	internal class InitInstance : MonoBehaviour {
+		private bool active;
+
 		public void Start() {
 			if(MechJeb.TypesLoaded) {
 				Logger.Info("Initializing MechJeb instance...");
-				Logger.Info(MechJeb.InitInstance() ? "KRPC.MechJeb is ready!" : "MechJeb found but the instance initialization wasn't successful. Maybe you don't have any MechJeb part attached to the vessel?");
+				Logger.Info((this.active = MechJeb.InitInstance()) ? "KRPC.MechJeb is ready!" : "MechJeb found but the instance initialization wasn't successful. Maybe you don't have any MechJeb part attached to the vessel?");
 				MechJeb.ShowErrors();
 			}
 		}
 
 		public void FixedUpdate() {
 			// Needed to fix outdated MechJeb instance when focus changes or a flight is reverted to launch.
-			if(FlightGlobals.ActiveVessel != MechJeb.Instance.vessel) {
+			if(this.active && FlightGlobals.ActiveVessel != MechJeb.Instance.vessel) {
 				Logger.Info("MechJeb instance changed, resetting cache...");
 				this.Start();
 			}